Barcode Analysis Using Linear Models

Tools for data management, count preprocessing, and differential analysis in massively parallel reporter assays (MPRA).

Details

This package provides tools for the analysis of MPRA data. The primary purpose is to enable powerful differential analysis of activity measures, but the package can also be used to generate precision weights useful in regression analyses of activity scores on sequence features. The main workhorse is the mpralm function which draws on the previously proposed voom framework for RNA-seq analysis in the limma package.

Author

Pia Keukeleire [cre, aut], Martin Kircher [aut]

Maintainer: Pia Keukeleire <pia.keukeleire@uksh.de>

References

Myint, Leslie, Dimitrios G. Avramopoulos, Loyal A. Goff, and Kasper D. Hansen. Linear models enable powerful differential activity analysis in massively parallel reporter assays. BMC Genomics 2019, 209. doi:10.1186/s12864-019-5556-x .

Law, Charity W., Yunshun Chen, Wei Shi, and Gordon K. Smyth. Voom: Precision Weights Unlock Linear Model Analysis Tools for RNA-Seq Read Counts. Genome Biology 2014, 15:R29. doi:10.1186/gb-2014-15-2-r29 .

Smyth, Gordon K., Joelle Michaud, and Hamish S. Scott. Use of within-Array Replicate Spots for Assessing Differential Expression in Microarray Experiments. Bioinformatics 2005, 21 (9): 2067-75. doi:10.1093/bioinformatics/bti270 .

Examples

data("mpraSetAggExample", package = "mpra")
design <- data.frame(intcpt = 1,
                     episomal = grepl("MT", colnames(mpraSetAggExample)))
mpralm_fit <- mpralm(object = mpraSetAggExample, design = design,
                     aggregate = "none", normalize = TRUE,
                     model_type = "indep_groups", plot = FALSE)
toptab <- topTable(mpralm_fit, coef = 2, number = Inf)
head(toptab)
#>                                                                                                                                                                                                                                                                                                                           logFC
#> R:FOXA1-NoMod_chr1:94430392-94430502_[chr1:94430361-94430532]                                                                                                                                                                                                                                                        -0.9037027
#> R:FOXA1_FOXA2-NoMod_chr3:171593815-171593944_[chr3:171593794-171593965]                                                                                                                                                                                                                                               0.8091496
#> C:SLEA_hg18:chr2:210861483-210861650|5:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;26:V_HNF1_C:AGTTAATGATTAACCAA;45:V_HNF1_C:AGTTAATGATTAACCAA;64:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;85:V_HNF1_C:AGTTAATGATTAACCAA;104:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;125:V_HNF1_C:AGTTAATGATTAACCAA;144:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC       -1.1786313
#> C:SLEA_hg18:chr2:210861483-210861650|11:V_HNF4_Q6:AAGGTCCAG;23:V_COUPTF_Q6:CCCCCTGACCTTTGCCCCCTGCC;49:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;71:V_HNF3ALPHA_Q6:TGTTTGCTTTG;85:V_Rxra_UP:GGCCGTGACCCCGTGAT;105:V_XBP1_01:GTGATGACGTGTCCCAT;125:V_HNF6_Q6:CAAAATCAATAA;140:V_HNF1_C:AGTTAATGATTAACCAA                        -1.1988885
#> C:SLEA_hg18:chr2:210861483-210861650|2:V_HNF1_C:AGTTAATGATTAACCAA;23:V_Rxra_UP:GGCCGTGACCCCGTGAT;44:V_Rxra_UP:GGCCGTGACCCCGTGAT;65:V_HNF1_C:AGTTAATGATTAACCAA;86:V_Rxra_UP:GGCCGTGACCCCGTGAT;107:V_Rxra_UP:GGCCGTGACCCCGTGAT;128:V_HNF1_C:AGTTAATGATTAACCAA;149:V_Rxra_UP:GGCCGTGACCCCGTGAT                          -0.7377998
#> C:SLEA_hg18:chr2:210861483-210861650|4:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;25:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;46:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;67:V_HNF1_C:AGTTAATGATTAACCAA;86:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;107:V_HNF1_C:AGTTAATGATTAACCAA;126:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;147:V_HNF1_C:AGTTAATGATTAACCAA -1.1865356
#>                                                                                                                                                                                                                                                                                                                         AveExpr
#> R:FOXA1-NoMod_chr1:94430392-94430502_[chr1:94430361-94430532]                                                                                                                                                                                                                                                         0.6371768
#> R:FOXA1_FOXA2-NoMod_chr3:171593815-171593944_[chr3:171593794-171593965]                                                                                                                                                                                                                                              -0.1902611
#> C:SLEA_hg18:chr2:210861483-210861650|5:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;26:V_HNF1_C:AGTTAATGATTAACCAA;45:V_HNF1_C:AGTTAATGATTAACCAA;64:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;85:V_HNF1_C:AGTTAATGATTAACCAA;104:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;125:V_HNF1_C:AGTTAATGATTAACCAA;144:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC        1.3427895
#> C:SLEA_hg18:chr2:210861483-210861650|11:V_HNF4_Q6:AAGGTCCAG;23:V_COUPTF_Q6:CCCCCTGACCTTTGCCCCCTGCC;49:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;71:V_HNF3ALPHA_Q6:TGTTTGCTTTG;85:V_Rxra_UP:GGCCGTGACCCCGTGAT;105:V_XBP1_01:GTGATGACGTGTCCCAT;125:V_HNF6_Q6:CAAAATCAATAA;140:V_HNF1_C:AGTTAATGATTAACCAA                         1.3967245
#> C:SLEA_hg18:chr2:210861483-210861650|2:V_HNF1_C:AGTTAATGATTAACCAA;23:V_Rxra_UP:GGCCGTGACCCCGTGAT;44:V_Rxra_UP:GGCCGTGACCCCGTGAT;65:V_HNF1_C:AGTTAATGATTAACCAA;86:V_Rxra_UP:GGCCGTGACCCCGTGAT;107:V_Rxra_UP:GGCCGTGACCCCGTGAT;128:V_HNF1_C:AGTTAATGATTAACCAA;149:V_Rxra_UP:GGCCGTGACCCCGTGAT                           1.0580753
#> C:SLEA_hg18:chr2:210861483-210861650|4:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;25:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;46:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;67:V_HNF1_C:AGTTAATGATTAACCAA;86:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;107:V_HNF1_C:AGTTAATGATTAACCAA;126:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;147:V_HNF1_C:AGTTAATGATTAACCAA  1.5333158
#>                                                                                                                                                                                                                                                                                                                              t
#> R:FOXA1-NoMod_chr1:94430392-94430502_[chr1:94430361-94430532]                                                                                                                                                                                                                                                        -17.61947
#> R:FOXA1_FOXA2-NoMod_chr3:171593815-171593944_[chr3:171593794-171593965]                                                                                                                                                                                                                                               17.21285
#> C:SLEA_hg18:chr2:210861483-210861650|5:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;26:V_HNF1_C:AGTTAATGATTAACCAA;45:V_HNF1_C:AGTTAATGATTAACCAA;64:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;85:V_HNF1_C:AGTTAATGATTAACCAA;104:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;125:V_HNF1_C:AGTTAATGATTAACCAA;144:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC       -16.75546
#> C:SLEA_hg18:chr2:210861483-210861650|11:V_HNF4_Q6:AAGGTCCAG;23:V_COUPTF_Q6:CCCCCTGACCTTTGCCCCCTGCC;49:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;71:V_HNF3ALPHA_Q6:TGTTTGCTTTG;85:V_Rxra_UP:GGCCGTGACCCCGTGAT;105:V_XBP1_01:GTGATGACGTGTCCCAT;125:V_HNF6_Q6:CAAAATCAATAA;140:V_HNF1_C:AGTTAATGATTAACCAA                        -16.01734
#> C:SLEA_hg18:chr2:210861483-210861650|2:V_HNF1_C:AGTTAATGATTAACCAA;23:V_Rxra_UP:GGCCGTGACCCCGTGAT;44:V_Rxra_UP:GGCCGTGACCCCGTGAT;65:V_HNF1_C:AGTTAATGATTAACCAA;86:V_Rxra_UP:GGCCGTGACCCCGTGAT;107:V_Rxra_UP:GGCCGTGACCCCGTGAT;128:V_HNF1_C:AGTTAATGATTAACCAA;149:V_Rxra_UP:GGCCGTGACCCCGTGAT                          -15.80748
#> C:SLEA_hg18:chr2:210861483-210861650|4:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;25:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;46:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;67:V_HNF1_C:AGTTAATGATTAACCAA;86:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;107:V_HNF1_C:AGTTAATGATTAACCAA;126:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;147:V_HNF1_C:AGTTAATGATTAACCAA -15.63195
#>                                                                                                                                                                                                                                                                                                                           P.Value
#> R:FOXA1-NoMod_chr1:94430392-94430502_[chr1:94430361-94430532]                                                                                                                                                                                                                                                        2.084956e-08
#> R:FOXA1_FOXA2-NoMod_chr3:171593815-171593944_[chr3:171593794-171593965]                                                                                                                                                                                                                                              2.570327e-08
#> C:SLEA_hg18:chr2:210861483-210861650|5:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;26:V_HNF1_C:AGTTAATGATTAACCAA;45:V_HNF1_C:AGTTAATGATTAACCAA;64:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;85:V_HNF1_C:AGTTAATGATTAACCAA;104:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;125:V_HNF1_C:AGTTAATGATTAACCAA;144:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC       3.271026e-08
#> C:SLEA_hg18:chr2:210861483-210861650|11:V_HNF4_Q6:AAGGTCCAG;23:V_COUPTF_Q6:CCCCCTGACCTTTGCCCCCTGCC;49:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;71:V_HNF3ALPHA_Q6:TGTTTGCTTTG;85:V_Rxra_UP:GGCCGTGACCCCGTGAT;105:V_XBP1_01:GTGATGACGTGTCCCAT;125:V_HNF6_Q6:CAAAATCAATAA;140:V_HNF1_C:AGTTAATGATTAACCAA                        4.891607e-08
#> C:SLEA_hg18:chr2:210861483-210861650|2:V_HNF1_C:AGTTAATGATTAACCAA;23:V_Rxra_UP:GGCCGTGACCCCGTGAT;44:V_Rxra_UP:GGCCGTGACCCCGTGAT;65:V_HNF1_C:AGTTAATGATTAACCAA;86:V_Rxra_UP:GGCCGTGACCCCGTGAT;107:V_Rxra_UP:GGCCGTGACCCCGTGAT;128:V_HNF1_C:AGTTAATGATTAACCAA;149:V_Rxra_UP:GGCCGTGACCCCGTGAT                          5.502017e-08
#> C:SLEA_hg18:chr2:210861483-210861650|4:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;25:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;46:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;67:V_HNF1_C:AGTTAATGATTAACCAA;86:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;107:V_HNF1_C:AGTTAATGATTAACCAA;126:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;147:V_HNF1_C:AGTTAATGATTAACCAA 6.077552e-08
#>                                                                                                                                                                                                                                                                                                                         adj.P.Val
#> R:FOXA1-NoMod_chr1:94430392-94430502_[chr1:94430361-94430532]                                                                                                                                                                                                                                                        1.667811e-05
#> R:FOXA1_FOXA2-NoMod_chr3:171593815-171593944_[chr3:171593794-171593965]                                                                                                                                                                                                                                              1.667811e-05
#> C:SLEA_hg18:chr2:210861483-210861650|5:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;26:V_HNF1_C:AGTTAATGATTAACCAA;45:V_HNF1_C:AGTTAATGATTAACCAA;64:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;85:V_HNF1_C:AGTTAATGATTAACCAA;104:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;125:V_HNF1_C:AGTTAATGATTAACCAA;144:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC       1.667811e-05
#> C:SLEA_hg18:chr2:210861483-210861650|11:V_HNF4_Q6:AAGGTCCAG;23:V_COUPTF_Q6:CCCCCTGACCTTTGCCCCCTGCC;49:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;71:V_HNF3ALPHA_Q6:TGTTTGCTTTG;85:V_Rxra_UP:GGCCGTGACCCCGTGAT;105:V_XBP1_01:GTGATGACGTGTCCCAT;125:V_HNF6_Q6:CAAAATCAATAA;140:V_HNF1_C:AGTTAATGATTAACCAA                        1.667811e-05
#> C:SLEA_hg18:chr2:210861483-210861650|2:V_HNF1_C:AGTTAATGATTAACCAA;23:V_Rxra_UP:GGCCGTGACCCCGTGAT;44:V_Rxra_UP:GGCCGTGACCCCGTGAT;65:V_HNF1_C:AGTTAATGATTAACCAA;86:V_Rxra_UP:GGCCGTGACCCCGTGAT;107:V_Rxra_UP:GGCCGTGACCCCGTGAT;128:V_HNF1_C:AGTTAATGATTAACCAA;149:V_Rxra_UP:GGCCGTGACCCCGTGAT                          1.667811e-05
#> C:SLEA_hg18:chr2:210861483-210861650|4:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;25:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;46:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;67:V_HNF1_C:AGTTAATGATTAACCAA;86:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;107:V_HNF1_C:AGTTAATGATTAACCAA;126:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;147:V_HNF1_C:AGTTAATGATTAACCAA 1.667811e-05
#>                                                                                                                                                                                                                                                                                                                              B
#> R:FOXA1-NoMod_chr1:94430392-94430502_[chr1:94430361-94430532]                                                                                                                                                                                                                                                        10.060963
#> R:FOXA1_FOXA2-NoMod_chr3:171593815-171593944_[chr3:171593794-171593965]                                                                                                                                                                                                                                               9.882373
#> C:SLEA_hg18:chr2:210861483-210861650|5:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;26:V_HNF1_C:AGTTAATGATTAACCAA;45:V_HNF1_C:AGTTAATGATTAACCAA;64:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;85:V_HNF1_C:AGTTAATGATTAACCAA;104:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;125:V_HNF1_C:AGTTAATGATTAACCAA;144:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC        9.651808
#> C:SLEA_hg18:chr2:210861483-210861650|11:V_HNF4_Q6:AAGGTCCAG;23:V_COUPTF_Q6:CCCCCTGACCTTTGCCCCCTGCC;49:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;71:V_HNF3ALPHA_Q6:TGTTTGCTTTG;85:V_Rxra_UP:GGCCGTGACCCCGTGAT;105:V_XBP1_01:GTGATGACGTGTCCCAT;125:V_HNF6_Q6:CAAAATCAATAA;140:V_HNF1_C:AGTTAATGATTAACCAA                         9.262579
#> C:SLEA_hg18:chr2:210861483-210861650|2:V_HNF1_C:AGTTAATGATTAACCAA;23:V_Rxra_UP:GGCCGTGACCCCGTGAT;44:V_Rxra_UP:GGCCGTGACCCCGTGAT;65:V_HNF1_C:AGTTAATGATTAACCAA;86:V_Rxra_UP:GGCCGTGACCCCGTGAT;107:V_Rxra_UP:GGCCGTGACCCCGTGAT;128:V_HNF1_C:AGTTAATGATTAACCAA;149:V_Rxra_UP:GGCCGTGACCCCGTGAT                           9.148315
#> C:SLEA_hg18:chr2:210861483-210861650|4:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;25:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;46:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;67:V_HNF1_C:AGTTAATGATTAACCAA;86:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;107:V_HNF1_C:AGTTAATGATTAACCAA;126:V_AHRARNT_02:GGGGATCGCGTGCCAGCCC;147:V_HNF1_C:AGTTAATGATTAACCAA  9.050762