Class "MPRASet"

A container for data from massively parallel reporter assays (MPRA). Builds on the SummarizedExperiment class.

## Constructor
MPRASet(label = new("character"), ...)

## Accessors
getRNA(object, aggregate = FALSE)
getDNA(object, aggregate = FALSE)
getBarcode(object)
getEid(object)
getEseq(object)

Arguments

object

A MPRASet object.

aggregate

A logical indicating if data should be aggregated to the element level (by summing across barcodes).

label

A character vector of labels to store with the elements. When provided, the labels are aligned to the element IDs.

...

Further arguments passed to mpra::MPRASet.

Slots

Slots are as described in a SummarizedExperiment. Of particular interest are colData which describes the phenotype data. The assay slot holds the assay data, with specific assay names RNA and DNA (accessed by getRNA and getDNA). Element and barcode data are accessible in the rowData slot. We have chosen to store barcode and element as character to be flexible, although they are often DNA sequences (and could therefore be considered DNAStringSet (from package Biostrings)).

Extends

Class "SummarizedExperiment", directly.

Value

The constructor function MPRASet returns an object of class "MPRASet".

Accessors

getDNA:

Gets the DNA channel data.

getRNA:

Gets the RNA channel data.

getBarcode:

Gets the barcode, if present.

getEid:

Gets the element ID

getEseq:

Gets the element sequence, if present.

See also

Examples

showClass("MPRASet")
#> Class "MPRASet" [package "mpra"]
#> 
#> Slots:
#>                                                                               
#> Name:            colData            assays             NAMES   elementMetadata
#> Class:         DataFrame    Assays_OR_NULL character_OR_NULL         DataFrame
#>                         
#> Name:           metadata
#> Class:              list
#> 
#> Extends: 
#> Class "SummarizedExperiment", directly
#> Class "RectangularData", by class "SummarizedExperiment", distance 2
#> Class "Vector", by class "SummarizedExperiment", distance 2
#> Class "Annotated", by class "SummarizedExperiment", distance 3
#> Class "vector_OR_Vector", by class "SummarizedExperiment", distance 3