Total count normalization of DNA and RNA counts

Total count normalization of DNA and RNA counts.

normalize_counts(object, normalizeSize = 1e+07, block = NULL)

Arguments

object

An object of class MPRASet.

normalizeSize

A numeric value specifying the total count to which all samples should be normalized. Default is 1e+07 (10 million).

block

A vector giving the sample designations of the columns of object. The default, NULL, indicates that all columns are separate samples.

Details

block is a vector that is used when the columns of the MPRAset object are paired. This often is the case when comparing allelic versions of an element. In this case, the first $s$ columns of object give the counts for the reference allele in $s$ samples. The second $s$ columns give the counts for the alternative allele measured in the same $s$ samples. With 3 samples, block would be c(1,2,3,1,2,3). All columns are scaled to have 10 million counts.

Value

An object of class MPRASet with the total count-normalized DNA and RNA counts.

Examples

data("mpraSetAggExample", package = "mpra")
mpraSetAggExample <- normalize_counts(mpraSetAggExample, )